Refactoring #1
No need to move mouse

getArtCenter method refactored

Extract IF-statement condition into WHILE-condition

is_in_between_rows method renamed into is_between_rows

Concatenate IF-statements

If the same code is repeated several times then make a method out of it

No need to make a click on the first artifact, waitSwitched method will take care of that

If total time is bigger OR EQUAL to max_wait then the function will return False

If ndigits (second argument) is omitted or is None, it returns the nearest integer to its input.

Do not use FOR-loop to scroll the page

Simplify IF-statements if possible

Let the FOR-loop control the flow

+++ ArtScanner/Tools/model_trainer/train_model.py
@@ -618,7 +618,7 @@ def detect_star(art_img):
     cropped_star = crop(normalize(to_gray(star)))
     coef = cropped_star.shape[1] / cropped_star.shape[0]
     coef = coef / 1.30882352 + 0.21568627
-    return int(round(coef))
+    return round(coef)
 
 
 filepath = "./train/weights-improvement-{epoch:02d}-{ctc_accu:.2f}.hdf5"
